refactor: modalController 리팩토링
